feat: arithemtic expressions?
precidence is messed up two-fold

1 - operator precedence is always to the right

meaning a + b + c + d == (((a + b) + c) + d)

This is a problem in the parser.

2 - parens work, but they don't

because of the way the code is generated, regs get clobbered quickly

for example 10 * (5 + 4) * 2 returns 72

why? the syntax tree looks like the following

      *
    /   \
   *     2
 /   \
10    +
    /   \
   5     4

the code generated looks like this
RAX = 10
RAX = 5
RCX = 4
RAX = RAX (5) + RCX (4)
RAX = RAX (9) * RCX (4)
RCX = 2
RAX = RAX (36) * RCX (2)

which gives 72...

without parens, life would be good because of problem 1
since the tree is always reparented, we get something that looks like this:

         *
       /   \
      *     d
    /   \
   +    c
 /   \
a     b

the code gen works specifically for trees like this, but we want it to
work for more than that

This is a problem in the code gen.

------

We can fix up both the parser and code gen.

Parser is relatively easy.

Code gen required some more thought. I have 2 options as I see it.
Either, for `n` operations, we allocate `n-1` space on the stack for the
results, then we compute the final operations by moving the final two
operands off the stack and into RAX. This will not reuse stack space
very well, but it would do the trick.

The other option is to keep track of which registers are in use as we
recurse through the arith tree and exit when its impossible to find more regs.

Yes, there is a third option where we do both, which has the added
benefit of being both super efficient with stack and regs AND allowing
for any arbitrary arith tree. However, I have no idea how to compute
the stack beforehand. If I can generate a function's code and then write
the code that sets up and tears down the stack, that would be ideal
because I can assume that I can get as much stack space as I want
while generating the code and then figure it out later.

We'll have to see.