feat: add division
since division requires the use of RAX and RDX, the best way to do it
with my current system is to disable those registers for general use. We
can maybe do a lot of cool optimization stuff if we switch to using an IR?

For now this will suffice in laying out the base compiler.