feat: function calls
I can't believe I overlooked the fact that my grammar doesn't
allow for function calls in expressions!! This gives me the perfect
opportunity to redesign my whole grammar with all that I have learned
in the last couple months.

Its no secret that dumlang is pretty much a subset of C which is a
little disappointing since one of the coolest parts of building a
compiler for a custom language is that its *your* syntax rules.
Hopefully I get it right this time, but even then I'll have more fun
modifying it later on than I do now.

That being said, its been hard to find time for this project in the last
couple weeks or so. This is partially due to a good chunk of time being
devoted to reading a book about parsing that a wikipedia article
recommended and that my school's library happened to have. Its been a
lot of fun to read, but definitely a big time sink. Ideally I can take
what I learn from this book and apply to my lexing and parsing
techniques since the way I do it now is... less than ideal. Other
factors include increasing academic and career requirements recently,
but I should be getting a handle on all of that very soon.

Thanks for reading! I hope to chat soon!