feat: new grammar definition
Finally got around to defining a new grammar, this time with some
example code!

I'm pretty happy with the new language. Its unique (enough) and isn't
annoying to write which were my two goals.

My next couple steps will be cleaning up the lexer and parser. The lexer
needs to be cleaned up yesterday. From what I've read, there are
"direct" and "indirect" lexers. The direct seems best for what we're
doing which is parsing all the tokens and then passing them to the
parser. However, when writing the parser I noticed two things I want
to implement: better error management and a less sequential compiler
pipeline. For the former, I need a way to "try parse" a certain
structure so that my parsing code wouldn't need to peek the next token
if two elements started with, for example, an identifier. Instead,
I'd "try parse" the first, if it fails "try parse" the second, and if
both fail then propagate the error. For the ladder, instead of parsing
alllllllllllllllllll the tokens, I want the parser's `next_token()`
function to call the lexer and read a little bit more of the file. The
challenge is getting these two to work together. We will need to do some
buffering because if a parse fails, we need to reset to the correct
'next token' without undoing all our hard work. I imagine a 'commit' at
the end of every successful parsed structure which clears the buffer and
does some other clean up. We will need to implement this sort of
buffering for reading from the file and reading the next token. To
bring this full circle, an indirect lexer might be more promising
because we often say "parse this specific token or error". Being able to
let the lexer know the token we want can improve performance a bit which
is a nice addition.

Overall, feeling excited to get these crucial components in a better state.