feat: begin work on better lexer
Really happy with how this is looking!

I removed some unnecessary tokens and that god awful parse_tokens
function. The lexer is much more readable AND its built with the
'try_parse' idea I talked about in the last commit: if it fails, no harm
just go back. The file seeking I assume is just as efficient as buffering a
block of the file, but I don't know enough to be sure.

At the very least, the implementation is clean enough to build off of
easily, and it also sets us up to build the `try_parse` functions in the
parser!