fix: implement try_parse functions correctly
Didn't take long to find that my trys weren't implemented correctly.

The best solution I could come up with was to have a fail label that
each ASSERT macro would jump to. The main reason is to reset the lexer
position on fail, but its flexible enough to free and cleanup any data
too.

A cool realization I made after rethinking my approach was that one
"stable position" variable is not enough when chaining a bunch of try
calls together. I imagine little markers in the string that you can
always jump back to. It was very reminiscent of the problem of recursive
function calls in assembly, how you have to push the return address onto
the stack if you're going to call a function. If you have a function
call chain of size N, you need to save the return address N-1 times, in
much the same way that if you need to parse N nested grammatical
structures, you need N-1 'stable positions'.