feat: finish parser rewrite!
So satisfying to see this thing work! I learned a lot along the way and
now we have a better lexer, parser, and grammar. Its just so awesome to
type out a simple program that calls some functions and manipulates
some numbers. The next steps are going to be system calls and then
hopefully a custom loader written in assembly that functions like
libcrt (maybe call it libdumrt?).

Had some rough edges with the direct lexer which allowed
some odd syntax, namely `dec my_var_name` and `decmy_var_name` are
equivalent because we only looked for `dec` and not `dec `. I've changed
it because there is no world where you can have *any* token directly
next to `dec`. However, there is an obvious problem when it comes to
other keywords, specifically types. We can't just say `int ` or ` int`
because `@func(v: int )` and `@func(v:int)` are both valid. Our direct
lexer needs to be a little more generic and parse alphanumeric keywords
that are surrounded my non-alphanumeric characters. I'm not sure if its
strictly necessary to check the preceding char, but its definitely
important to check the following one. I will fix this issue at some
point.