feat: while loop!!!
Implemented a little program that calculates how many digits in a number
in the example which was SO satisfying to see work!

A frequent problem I'm running into is having to 'rewrite' jump
instructions after finding out the relative distance they need. I need
a cleaner way that doesn't require saving the instruction location (for
overwriting purposes), the location after the instruction (since
relative distance is always relative to the position AFTER the jump
instruction), and the location to jump back to after overwriting. Maybe
a function that takes a funcptr, an int32_t for the jump offset, and a
size_t for the location in the text block will do for the time being? I
feel like there might be a better way, but the fact remains that
we *must* save the position of the instruction. Maybe hard coding the
j and je instruction lengths isn't so bad? Maybe we can leverage scope a
little better?

I'll consider these thoughts more in the future. For now, I'm very happy.