fix: scope local jmptab
Probably should have seen this coming lol

The problem was when you nest conditions like the following:

while A {
  if B {
    // X
  }
}

The jmptab has LABEL_BLOCK_START and LABEL_BLOCK_END entries in it after
evaluating the first condition (A) and, to compute where those are, we
have to write the code block containing another condition (B). By the
time we evaluate condition B, we have 2 LABEL_BLOCK_STARTs and
LABEL_BLOCK_ENDs that are intended to be different, so after we write
the code block, we evaluate all of the labels to the start and end of
B's code block.

This could just be The Lord speaking to me, but I'm getting the
sense that global isn't very cohesive to all the recursion I'm doing.