fix: ret jumps, fix ABI compat (mostly)
jmptab_t is now a doubly linked list so that we can remove easily. All
non-resolved labels are merged into the super-scope's jmptab. This has
made return possible, and will hopefully make `continue` and `break`
also easy to implement.

I wanted to test if the while loops and if statements were behaving as
they should so I wrote an equivalent function in C to test. I called the
functions in the params of printf and they weren't lining up. I was
baffled for a minute, not sure why *C* was the wrong one. I changed it
so that I would save the results of these functions in variables first,
and low and behold it worked, so I figured I was breaking ABI somehow.
This commit pushes the preserved registers on the stack and pops them
off at the end, but there is still ABI incompatability. As of right now,
you can modify parameter variables. I plan on adding a flag to the
scope_var_t struct that indicates its immutable. I don't think I want to
implement a way to be able to modify these variables without breaking
ABI. Just create a local variable and copy the value to that; that's
what I'd effectively have the compiler do anyways.

Fun stuff!